Renamed the sample implementation and fixed a bug
